EmfModel: split AbstractReflectiveEmfModel off
