+++ plugins/org.eclipse.epsilon.emc.emf/src/org/eclipse/epsilon/emc/emf/EmfModel.java
@@ -11,19 +11,21 @@ package org.eclipse.epsilon.emc.emf;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
 import org.eclipse.emf.common.notify.Adapter;
 import org.eclipse.emf.common.notify.Notification;
-import org.eclipse.emf.common.util.EList;
-import org.eclipse.emf.common.util.Enumerator;
 import org.eclipse.emf.common.util.URI;
-import org.eclipse.emf.ecore.EEnumLiteral;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EPackage;
 import org.eclipse.emf.ecore.EReference;
-import org.eclipse.emf.ecore.EStructuralFeature;
 import org.eclipse.emf.ecore.EcorePackage;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
@@ -34,13 +36,10 @@ import org.eclipse.epsilon.common.util.CollectionUtil;
 import org.eclipse.epsilon.common.util.StringProperties;
 import org.eclipse.epsilon.eol.exceptions.models.EolModelElementTypeNotFoundException;
 import org.eclipse.epsilon.eol.exceptions.models.EolModelLoadingException;
-import org.eclipse.epsilon.eol.exceptions.models.EolNotAnEnumerationValueException;
-import org.eclipse.epsilon.eol.execute.introspection.IReflectivePropertySetter;
 import org.eclipse.epsilon.eol.m3.Metamodel;
-import org.eclipse.epsilon.eol.models.IReflectiveModel;
 import org.eclipse.epsilon.eol.models.IRelativePathResolver;
 
-public class EmfModel extends AbstractEmfModel implements IReflectiveModel {
+public class EmfModel extends AbstractReflectiveEmfModel {
 
 	/**
 	 * @deprecated {@link #PROPERTY_METAMODEL_URI} and {@link #PROPERTY_FILE_BASED_METAMODEL_URI} are now
@@ -131,33 +130,8 @@ public class EmfModel extends AbstractEmfModel implements IReflectiveModel {
 	
 
 	public EmfModel() {
+		super();
 		propertySetter = new EmfPropertySetter();
-	}
-	
-	@Override
-	public IReflectivePropertySetter getPropertySetter() {
-		return (IReflectivePropertySetter) propertySetter;
-	}
-	
-	@Override
-	public Collection<String> getPropertiesOf(String type) throws EolModelElementTypeNotFoundException {
-		Collection<EStructuralFeature> features = featuresForType(type);
-		final Collection<String> properties = new ArrayList<>(features.size());
-		for (EStructuralFeature feature : features) {
-			properties.add(feature.getName());
-		}
-		return properties;
-	}
-
-	@Override
-	public boolean preventLoadingOfExternalModelElements() {
-		if (isExpand()) {
-			setExpand(false);
-			return true;
-		}
-		else {
-			return false;
-		}
 	}
 	
 	/**
@@ -525,36 +499,6 @@ public class EmfModel extends AbstractEmfModel implements IReflectiveModel {
 	}
 
 	@Override
-	public boolean hasProperty(String type, String property) throws EolModelElementTypeNotFoundException {
-		return getPropertiesOf(type).contains(property);
-	}
-	
-	@Override
-	public boolean isEnumerationValue(Object object) {
-		return object instanceof Enumerator;
-	}
-
-	@Override
-	public String getEnumerationTypeOf(Object literal) throws EolNotAnEnumerationValueException {
-		if (!isEnumerationValue(literal))
-			throw new EolNotAnEnumerationValueException(literal);
-		
-		if (literal instanceof EEnumLiteral) {
-			return ((EEnumLiteral)literal).getEEnum().getName();
-		} else {
-			return ((Enumerator)literal).getClass().getSimpleName();
-		}		
-	}
-
-	@Override
-	public String getEnumerationLabelOf(Object literal) throws EolNotAnEnumerationValueException {
-		if (!isEnumerationValue(literal))
-			throw new EolNotAnEnumerationValueException(literal);
-			
-		return ((Enumerator)literal).getName();
-	}
-
-	@Override
 	public String toString() {
 		return "EmfModel [name=" + getName() + "]";
 	}
@@ -625,48 +569,6 @@ public class EmfModel extends AbstractEmfModel implements IReflectiveModel {
 		final Long timestamp = metamodelFile.lastModified();
 		fileBasedMetamodels.put(path, packages);
 		fileBasedMetamodelTimestamps.put(path, timestamp);
-	}
-
-	@Override
-	public boolean hasPackage(String packageName) {
-		return packageForName(packageName) != null;
-	}
-	
-	private EPackage packageForName(String name) {
-		final String[] parts = name.split("::");
-		
-		int partIndex = 0;
-		EPackage current;
-		Collection<EPackage> next = getTopLevelPackages();
-		
-		do {
-			if ((current = packageForName(parts[partIndex++], next)) != null) {
-				next = current.getESubpackages();
-			}
-		}
-		while (current != null && partIndex < parts.length);
-		
-		return current;
-	}
-
-	private Collection<EPackage> getTopLevelPackages() {
-		return getPackageRegistry().values()
-			.stream()
-			.filter(pkg -> pkg instanceof EPackage)
-			.map(pkg -> (EPackage) pkg)
-			.collect(Collectors.toCollection(LinkedList::new));
-	}
-	
-	private static EPackage packageForName(String name, Collection<EPackage> packages) {
-		return packages.stream()
-			.filter(pkg -> name.equals(pkg.getName()))
-			.findAny()
-			.orElse(null);
-	}
-	
-
-	private EList<EStructuralFeature> featuresForType(String type) throws EolModelElementTypeNotFoundException {
-		return classForName(type).getEAllStructuralFeatures();
 	}
 
 	private static List<URI> toURIList(final String commaSeparatedList) {
